% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-NA.R
\docType{data}
\name{NA_x_}
\alias{NA_x_}
\alias{NA_Date_}
\alias{NA_POSIXct_}
\title{Not Available / Missing Values}
\format{
An object of class \code{Date} of length 1.

An object of class \code{POSIXct} (inherits from \code{POSIXt}) of length 1.
}
\usage{
NA_Date_

NA_POSIXct_
}
\value{
\code{NA} value of a particular type.
}
\description{
\code{NA} is a logical constant of length 1 which contains a missing value
indicator. Other built-in constants \code{NA_integer_}, \code{NA_real_},
\code{NA_complex_} and \code{NA_character_} are missing value place-holder of
a particular type.

NA constants have two typical uses:
\enumerate{
\item Signifying the user what data type a function expect in its input; and
\item Representing missing values of a specific type.
}

This package provides additional \code{NA} types.
}
\examples{
class(NA_Date_)
class(NA_POSIXct_)
}
\keyword{datasets}
