% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\alias{\%||\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs

x \%||\% y
}
\value{
The result of \code{rhs(lfs)}.

If \code{rhs} is NULL, will return \code{lhs}; otherwise returns \code{rhs}.
}
\description{
Pipe operator

NULL operator
}
\keyword{internal}
